+++ src/main/java/com/merge/final_project/campaign/settlement/service/SettlementCommandService.java
@@ -4,7 +4,7 @@ import com.merge.final_project.campaign.campaigns.entity.Campaign;
 import com.merge.final_project.campaign.settlement.Repository.SettlementRepository;
 import com.merge.final_project.campaign.settlement.Settlement;
 import com.merge.final_project.campaign.settlement.SettlementStatus;
-import com.merge.final_project.org.foundation.Foundation;
+import com.merge.final_project.org.Foundation;
 import com.merge.final_project.recipient.beneficiary.Beneficiary;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,16 +19,15 @@ public class SettlementCommandService {
 
     private final SettlementRepository settlementRepository;
 
-    // 정산 원본은 체인 처리와 분리해서 먼저 저장해야
-    // 체인 호출 실패가 나도 어떤 정산 시도가 실패했는지 남길 수 있다.
+    // 정산 생성 (초기 상태: PENDING)
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public Settlement createPendingSettlement(
             String transactionCode,
             Foundation foundation,
             Beneficiary beneficiary,
-            Integer totalAmount,
-            Integer foundationAmount,
-            Integer beneficiaryAmount,
+            Long totalAmount,
+            Long foundationAmount,
+            Long beneficiaryAmount,
             Campaign campaign
     ) {
         Settlement settlement = Settlement.builder()
@@ -44,9 +43,7 @@ public class SettlementCommandService {
 
         return settlementRepository.save(settlement);
     }
-
-    // 실제 온체인 정산을 시작하기 직전에 PROCESSING 으로 바꿔두면,
-    // 체인 성공 후 로컬 저장만 실패한 건을 PENDING 과 구분해서 볼 수 있다.
+    // 처리 시작 (PENDING → PROCESSING)
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void markProcessing(Long settlementNo) {
         Settlement settlement = settlementRepository.findById(settlementNo)
@@ -54,7 +51,7 @@ public class SettlementCommandService {
 
         settlement.markProcessing();
     }
-
+    // 정산 완료 처리 (→ COMPLETED)
     @Transactional
     public void markCompleted(Long settlementNo) {
         Settlement settlement = settlementRepository.findById(settlementNo)
@@ -62,7 +59,7 @@ public class SettlementCommandService {
         settlement.setStatus(SettlementStatus.COMPLETED);
         settlement.setSettledAt(LocalDateTime.now());
     }
-
+    // 정산 실패 처리 (→ FAILED)
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void markFailed(Long settlementNo) {
         Settlement settlement = settlementRepository.findById(settlementNo)
